Added restart feature, a spiffy progressbar, and the ability to choose power-law radiation sources.

+++ src/rt1d.py
@@ -9,8 +9,9 @@ Description: Driver script for a 1D radiative transfer code.
 
 """
 
-import sys, os
+import sys, os, time
 import numpy as np
+from progressbar import *
 
 try:
     from mpi4py import MPI
@@ -27,38 +28,59 @@ from InitializeParameterSpace import *
 from InitializeGrid import *
 from Radiate import *
 from WriteData import *
+from ReadRestartFile import *
 
-# Retrieve name of parameter file from command line.
-pf = sys.argv[1]
+all_pfs = {}
 
-# Instantiate initialization class and build parameter space.
-ps = InitializeParameterSpace(pf)
-all_pfs = ps.AllParameterSets()
-del pf
+# Currently we only understand parameter files and the '-r' restart flag.  
+if sys.argv[1] == '-r':
+    IsRestart = True
+    rf = "{0}/{1}".format(os.getcwd(), sys.argv[2])
+    pf, ICs = ReadRestartFile(rf)
+    all_pfs["{0}0000".format(pf["SavePrefix"])] = pf
+    del pf
+
+else:
+    IsRestart = False
+    pf = sys.argv[1]
+
+    # Instantiate initialization class and build parameter space.
+    ps = InitializeParameterSpace(pf)
+    all_pfs = ps.AllParameterSets()
+    del pf
 
 if rank == 0: 
     print "\nStarting rt1d..."
     print "Initializing {0} 1D radiative transfer calculation(s)...".format(len(all_pfs)) 
+    if IsRestart: print "Restarting from {0}".format(rf)
     print "Press ctrl-C to quit at any time.\n" 
 
 # Loop over parameter sets. 
 for i, pf in enumerate(all_pfs):
     if i % size != rank: continue
-    this_pf = all_pfs[pf]
     
+    start = time.time()
+    
+    this_pf = all_pfs[pf]
+    this_pf["BaseName"] = pf
     TimeUnits = this_pf["TimeUnits"]
     StopTime = this_pf["StopTime"] * TimeUnits
     dt = this_pf["InitialTimestep"] * TimeUnits
     dtDataDump = this_pf["dtDataDump"] * TimeUnits
-        
-    try: os.mkdir("{0}".format(pf))
-    except OSError: 
-        os.system("rm -rf {0}".format(pf))
-        os.mkdir("{0}".format(pf))
     
-    # Initialize grid
-    g = InitializeGrid(this_pf)   
-    data = g.InitializeFields()
+    # Widget for progressbar.
+    widget = ["rt1d: ", Percentage(), ' ', Bar(marker = RotatingMarker()), ' ', ETA(), ' ']
+            
+    # Initialize grid and file system
+    if IsRestart: data = ICs
+    else:
+        g = InitializeGrid(this_pf)   
+        data = g.InitializeFields()
+        
+        try: os.mkdir("{0}".format(pf))
+        except OSError: 
+            os.system("rm -rf {0}".format(pf))
+            os.mkdir("{0}".format(pf))
 
     # Initialize integral tables
     iits = InitializeIntegralTables(this_pf, data)
@@ -68,19 +90,29 @@ for i, pf in enumerate(all_pfs):
     r = Radiate(this_pf, itabs, [iits.HIColumn, iits.HeIColumn, iits.HeIIColumn])
     w = WriteData(this_pf)
     
-    # Figure out data dump times, write out initial dataset.
+    # Figure out data dump times, write out initial dataset (or not if this is a restart).
     ddt = np.arange(0, StopTime + dtDataDump, dtDataDump)
-    w.WriteAllData(data, pf, 0, 0)
+    t = this_pf["CurrentTime"]
+    wct = int(t / dtDataDump) + 1
+    if not IsRestart: w.WriteAllData(data, 0, t)
         
-    t = 0.0
-    wct = 1
     while t <= StopTime:
+        
+        # Progress bar
+        if rank == 0:
+            if t == 0: print "rt1d:  t = 0.0 / {0}".format(StopTime / TimeUnits)
+            pbar = ProgressBar(widgets = widget, maxval = dtDataDump / TimeUnits).start()
+            pbar.update((t / TimeUnits) - ((wct - 1) * (dtDataDump / TimeUnits)))
+        
+        # Evolve photons
         data, dt = r.EvolvePhotons(data, t, dt)
-                        
-        print "t = {0}".format(t / TimeUnits)                
-                        
+               
+        # Write-out data, or don't                                        
         if t == ddt[wct]:
-            w.WriteAllData(data, pf, wct, t)
+            w.WriteAllData(data, wct, t)
+            if rank == 0: 
+                pbar.finish()
+                print "rt1d:  t = {0} / {1}".format(t / TimeUnits, StopTime / TimeUnits)
             wct += 1
         elif (t + dt) > ddt[wct]:
             dt = ddt[wct] - t
@@ -89,8 +121,9 @@ for i, pf in enumerate(all_pfs):
 
         t += dt
         
+    elapsed = time.time() - start    
     del g, r, w, data
-    print "Calculation {0} ({1}) complete.".format(i + 1, pf)
+    print "Calculation {0} ({1}) complete.  Elapsed time = {2} seconds.".format(i + 1, pf, round(elapsed, 2))
